+++ crawler/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+    "singleQuote": false,
+    "trailingComma": "all",
+    "printWidth": 110,
+    "tabWidth": 2
+  }
+
