+++ crawler/src/crawl/community.ts
@@ -2,7 +2,11 @@ import logging from "../lib/logging";
 
 import { CrawlError } from "../lib/error";
 
-import storage from "../crawlStorage";
+import { CrawlTooRecentError } from "../lib/error";
+
+import { IJobProcessor } from "../queue/BaseQueue";
+
+import crawlStorage from "../lib/crawlStorage";
 
 import CrawlClient from "../lib/CrawlClient";
 
@@ -43,83 +47,78 @@ export default class CommunityCrawler {
 
   // validate the community is for the domain being scanned, and save it
   async storeCommunityData(community) {
-    const { basePart, communityPart } = this.splitCommunityActorParts(
-      community.community.actor_id
-    );
+    const { basePart, communityPart } = this.splitCommunityActorParts(community.community.actor_id);
 
     // validate the community actor_id matches the domain
-    if (
-      basePart != this.crawlDomain ||
-      communityPart != community.community.name
-    ) {
+    if (basePart != this.crawlDomain || communityPart != community.community.name) {
       logging.error(
-        `${this.logPrefix} Community actor_id does not match domain: ${community.community.actor_id} ${community.community.name}`
+        `${this.logPrefix} Community actor_id does not match domain: ${community.community.actor_id} ${community.community.name}`,
       );
       return false;
     }
 
-    await storage.community.setTrackedAttribute(
+    await crawlStorage.community.setTrackedAttribute(
       this.crawlDomain,
       communityPart,
       "subscribers",
-      community.counts.subscribers
+      community.counts.subscribers,
     );
 
     if (community.counts.hot_rank) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "hot_rank",
-        community.counts.hot_rank
+        community.counts.hot_rank,
       );
     }
 
     if (community.counts.posts) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "posts",
-        community.counts.posts
+        community.counts.posts,
       );
     }
 
     if (community.counts.comments) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "comments",
-        community.counts.comments
+        community.counts.comments,
       );
     }
 
     if (community.counts.users_active_day) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "users_active_day",
-        community.counts.users_active_day
+        community.counts.users_active_day,
       );
     }
 
     if (community.counts.users_active_week) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "users_active_week",
-        community.counts.users_active_week
+        community.counts.users_active_week,
       );
     }
 
     if (community.counts.users_active_month) {
-      await storage.community.setTrackedAttribute(
+      await crawlStorage.community.setTrackedAttribute(
         this.crawlDomain,
         communityPart,
         "users_active_month",
-        community.counts.users_active_month
+        community.counts.users_active_month,
       );
     }
 
-    await storage.community.upsert(this.crawlDomain, community);
+    await crawlStorage.community.upsert(this.crawlDomain, community);
 
     return true;
   }
@@ -132,10 +131,7 @@ export default class CommunityCrawler {
 
       logging.info(`${this.logPrefix} Ended Success: ${communityName}`);
     } catch (e) {
-      logging.trace(
-        `${this.logPrefix} ERROR Community: ${communityName}`,
-        e.message
-      );
+      logging.trace(`${this.logPrefix} ERROR Community: ${communityName}`, e.message);
     }
   }
 
@@ -148,14 +144,11 @@ export default class CommunityCrawler {
             name: communityName,
           },
         },
-        0
+        0,
       );
 
       if (communityData.data.community_view) {
-        console.log(
-          `${this.logPrefix} Storing`,
-          communityData.data.community_view.community.name
-        );
+        console.log(`${this.logPrefix} Storing`, communityData.data.community_view.community.name);
 
         await this.storeCommunityData(communityData.data.community_view);
 
@@ -166,17 +159,13 @@ export default class CommunityCrawler {
       if (e.data && e.data.error == "couldnt_find_community") {
         logging.warn("DELETE community error", e.data.error);
 
-        await storage.community.delete(
-          this.crawlDomain,
-          communityName,
-          e.data.error
-        );
+        await crawlStorage.community.delete(this.crawlDomain, communityName, e.data.error);
         return;
       } else if (e.data) {
         logging.warn("OTHER community error");
 
         // only delete unless explicit
-        // await storage.community.delete(
+        // await crawlStorage.community.delete(
         //   this.crawlDomain,
         //   communityName,
         //   e.data.error
@@ -186,9 +175,7 @@ export default class CommunityCrawler {
 
       // re-try overall
       if (attempt < RETRY_COUNT) {
-        await new Promise((resolve) =>
-          setTimeout(resolve, TIME_BETWEEN_RETRIES)
-        );
+        await new Promise((resolve) => setTimeout(resolve, TIME_BETWEEN_RETRIES));
         return await this.getSingleCommunityData(communityName, attempt + 1);
       }
 
@@ -206,9 +193,7 @@ export default class CommunityCrawler {
       const promisesArray = await this.crawlCommunityPaginatedList();
       const resultPromises = await Promise.all(promisesArray);
 
-      logging.info(
-        `${this.logPrefix} Ended Success (${resultPromises.length} results)`
-      );
+      logging.info(`${this.logPrefix} Ended Success (${resultPromises.length} results)`);
 
       return resultPromises;
     } catch (e) {
@@ -221,9 +206,7 @@ export default class CommunityCrawler {
   async crawlCommunityPaginatedList(pageNumber: number = 1) {
     const communities = await this.getPageData(pageNumber);
 
-    logging.debug(
-      `${this.logPrefix} Page ${pageNumber}, Results: ${communities.length}`
-    );
+    logging.debug(`${this.logPrefix} Page ${pageNumber}, Results: ${communities.length}`);
 
     //  promises track the upsert of community data
     let promises: Promise<any>[] = [];
@@ -237,9 +220,7 @@ export default class CommunityCrawler {
       // sleep between pages
       await new Promise((resolve) => setTimeout(resolve, TIME_BETWEEN_PAGES));
 
-      const subPromises = await this.crawlCommunityPaginatedList(
-        pageNumber + 1
-      );
+      const subPromises = await this.crawlCommunityPaginatedList(pageNumber + 1);
       if (subPromises.length > 0) {
         promises.push(...subPromises);
       }
@@ -264,7 +245,7 @@ export default class CommunityCrawler {
           },
           timeout: PAGE_TIMEOUT,
         },
-        RETRY_PAGE_COUNT // retry count per-page
+        RETRY_PAGE_COUNT, // retry count per-page
       );
     } catch (e) {
       // throw new CrawlError("Failed to get community page");
@@ -282,3 +263,92 @@ export default class CommunityCrawler {
     return communities;
   }
 }
+
+export const communityListProcessor: IJobProcessor = async ({ baseUrl }) => {
+  const startTime = Date.now();
+
+  try {
+    // check if community's instance has already been crawled revcently (these expire from redis)
+    const lastCrawl = await crawlStorage.tracking.getLastCrawl("community", baseUrl);
+    if (lastCrawl) {
+      const lastCrawledMsAgo = Date.now() - lastCrawl.time;
+      throw new CrawlTooRecentError(`Skipping - Crawled too recently (${lastCrawledMsAgo / 1000}s ago)`);
+    }
+
+    // check when the latest entry to errors was too recent
+    const lastErrorTs = await crawlStorage.tracking.getOneError("community", baseUrl);
+    if (lastErrorTs) {
+      const lastErrorMsAgo = Date.now() - lastErrorTs.time;
+      throw new CrawlTooRecentError(`Skipping - Error too recently (${lastErrorMsAgo / 1000}s ago)`);
+    }
+
+    // perform the crawl
+    const crawler = new CommunityCrawler(baseUrl);
+    const communityData = await crawler.crawlList();
+
+    // set last successful crawl
+    await crawlStorage.tracking.setLastCrawl("community", baseUrl, {
+      duration: (Date.now() - startTime) / 1000,
+      resultCount: communityData.length,
+    });
+
+    const endTime = Date.now();
+    logging.info(`[Community] [${baseUrl}] Finished in ${(endTime - startTime) / 1000}s`);
+
+    return communityData;
+  } catch (error) {
+    if (error instanceof CrawlTooRecentError) {
+      logging.warn(`[Community] [${baseUrl}] CrawlTooRecentError: ${error.message}`);
+    } else {
+      const errorDetail = {
+        error: error.message,
+        stack: error.stack,
+        isAxiosError: error.isAxiosError,
+        requestUrl: error.isAxiosError ? error.request.url : null,
+        time: Date.now(),
+        duration: Date.now() - startTime,
+      };
+
+      logging.error(`[Community] [${baseUrl}] Error: ${error.message}`);
+
+      // if (error instanceof CrawlError || error instanceof AxiosError) {
+      await crawlStorage.tracking.upsertError("community", baseUrl, errorDetail);
+    }
+  }
+
+  return null;
+};
+
+export const singleCommunityProcessor: IJobProcessor = async ({ baseUrl, community }) => {
+  let communityData: any = null;
+
+  if (!baseUrl || !community) {
+    logging.error(`[OneCommunity] [${baseUrl}] Missing baseUrl or community`);
+    return null;
+  }
+
+  try {
+    const crawler = new CommunityCrawler(baseUrl);
+
+    communityData = await crawler.crawlSingle(community);
+  } catch (error) {
+    if (error instanceof CrawlTooRecentError) {
+      logging.warn(`[OneCommunity] [${baseUrl}] CrawlTooRecentError: ${error.message}`);
+    } else {
+      const errorDetail = {
+        error: error.message,
+        stack: error.stack,
+        isAxiosError: error.isAxiosError,
+        requestUrl: error.isAxiosError ? error.request.url : null,
+        time: Date.now(),
+      };
+
+      // if (error instanceof CrawlError || error instanceof AxiosError) {
+      await crawlStorage.tracking.upsertError("one_community", baseUrl, errorDetail);
+
+      logging.error(`[OneCommunity] [${baseUrl}] Error: ${error.message}`);
+    }
+  }
+
+  return communityData;
+};
