+++ crawler/src/crawl/uptime.ts
@@ -4,7 +4,7 @@
  */
 import logging from "../lib/logging";
 
-import crawlStorage from "../crawlStorage";
+import crawlStorage from "../lib/crawlStorage";
 
 import CrawlClient from "../lib/CrawlClient";
